Script to generate data file and claim
