+++ .github/workflows/create-claim.yml
@@ -1,0 +1,31 @@
+on:
+  push:
+    branches: [ main ]
+
+name: Create Claim
+
+jobs:
+  create-claim:
+    name: Create Claim
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Node Setup
+        uses: actions/setup-node@v1
+        with:
+          node-version: 12
+
+      - name: Install node dependencies
+        run: npm install
+
+      - name: Automated Create Claim
+        env:
+          NODE_URL: ${{ secrets.NODE_URL }}
+          MNEMONIC: ${{ secrets.MNEMONIC }}
+          PULL_REQUEST_TITLE : ${{ github.event.pull_request.title }}
+        run: npm run automated-create-claim
+          
+
