+++ .github/workflows/data-file-verification.yml
@@ -1,0 +1,28 @@
+name: Data file verification
+on: [pull_request]
+
+jobs:
+  data-file-verification:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node: [12]
+    steps:
+      - uses: actions/checkout@v2.3.1
+        with:
+          submodules: true
+      
+      - name: Use Node.js
+        uses: actions/setup-node@v1
+        with:
+          node-version: ${{ matrix.node }}
+
+      - name: Run npm install with node v${{ matrix.node }}
+        run: npm install
+
+      - name: Validate data file.
+        env:
+          NODE_URL: ${{ secrets.NODE_URL }}
+          PULL_REQUEST_TITLE : ${{ github.event.pull_request.title }}
+        run: npm run validate-data-file
+
