+++ package.json
@@ -13,9 +13,13 @@
     "lint:fix": "eslint --cache --fix .",
     "postinstall": "lerna bootstrap",
     "snapshot": "node index.js",
-    "create-claim": "node create-claim.js"
+    "create-claim": "node create-claim.js",
+    "create-data-file": "scripts/create-data-file.sh",
+    "validate-data-file": "node data-file-validator.js",
+    "automated-create-claim": "scripts/create-claim.sh"
   },
   "dependencies": {
+    "@actions/core": "^1.4.0",
     "@truffle/hdwallet-provider": "^1.2.0",
     "bip39": "^3.0.3",
     "dotenv": "^8.2.0",
